b/161479450: verify transcoding Any type
