+++ tests/endpoints/bookstore_grpc/client/client.go
@@ -117,6 +117,7 @@ var makeHTTPCall = func(addr, httpMethod, method, token string, header http.Head
 	return string(content), nil
 }
 
+//TODO(b/162626126): cleanup duplicate call methods.
 func makeHTTP2Call(addr, httpMethod, method, token string, header http.Header) (string, error) {
 	cli := http.Client{
 		// Skip TLS dial
