+++ tests/integration_test/transcoding_bindings_test.go
@@ -128,6 +128,14 @@ func TestTranscodingBindings(t *testing.T) {
 			method:         "/v1/shelves/100/books?key=api-key",
 			wantResp:       `{"books":[{"id":"1001","title":"Alphabet"},{"id":"4","author":"Leo Tolstoy","title":"War and Peace"},{"id":"5","author":"Mark","title":"The Adventures of Huckleberry Finn"},{"id":"6","author":"Foo/Bar/Baz","title":"The Adventures of Huckleberry Finn"}]}`,
 		},
+		{
+			desc:           "Succeed, test transcoding Any type json <-> grpc by echoing book in Any",
+			clientProtocol: "http",
+			httpMethod:     "POST",
+			method:         "/v1/shelves?key=api-key",
+			bodyBytes:      []byte(`{"id":"300","theme":"Horror","any":{"@type":"type.googleapis.com/endpoints.examples.bookstore.Book","id":"123","author":"author","title":"title"}}`),
+			wantResp:       `{"id":"300","theme":"Horror","any":{"@type":"type.googleapis.com/endpoints.examples.bookstore.Book","id":"123","author":"author","title":"title"}}`,
+		},
 	}
 	for _, tc := range tests {
 		addr := fmt.Sprintf("localhost:%v", s.Ports().ListenerPort)
@@ -139,192 +147,5 @@ func TestTranscodingBindings(t *testing.T) {
 				t.Errorf("Test (%s): failed, expected: %s, got: %s", tc.desc, tc.wantResp, resp)
 			}
 		}
-	}
-}
-
-func TestTranscodingPrintOptions(t *testing.T) {
-	t.Parallel()
-
-	configID := "test-config-id"
-	type testType struct {
-		desc                                  string
-		clientProtocol                        string
-		httpMethod                            string
-		method                                string
-		bodyBytes                             []byte
-		wantResp                              string
-		transcodingAlwaysPrintPrimitiveFields bool
-		transcodingAlwaysPrintEnumsAsInts     bool
-		transcodingPreserveProtoFieldNames    bool
-	}
-	tests := []testType{
-		{
-			desc:           "Success. Default setting used to be compared with other test cases.",
-			clientProtocol: "http",
-			httpMethod:     "POST",
-			method:         "/v1/shelves/100/books?key=api-key",
-			bodyBytes:      []byte(`{"id": 4, "type": 1, "author":"Mark", "price_in_usd": 100}`),
-			wantResp:       `{"id":"4","author":"Mark","type":"COMIC","price_in_usd":100}`,
-		},
-		{
-			desc:                                  "Success. Set transcoding_always_print_primitive_fields to true",
-			clientProtocol:                        "http",
-			httpMethod:                            "POST",
-			method:                                "/v1/shelves/100/books?key=api-key",
-			bodyBytes:                             []byte(`{"id": 4}`),
-			transcodingAlwaysPrintPrimitiveFields: true,
-			wantResp:                              `{"id":"4","author":"","title":"","type":"CLASSIC","price_in_usd":0}`,
-		},
-		{
-			desc:                              "Success. Set transcoding_always_print_enums_as_ints to true",
-			clientProtocol:                    "http",
-			httpMethod:                        "POST",
-			method:                            "/v1/shelves/100/books?key=api-key",
-			bodyBytes:                         []byte(`{"id": 4, "type":1}`),
-			transcodingAlwaysPrintEnumsAsInts: true,
-			wantResp:                          `{"id":"4","type":1}`,
-		},
-		{
-			desc:                               "Success. Set transcoding_preserve_proto_field_names to true",
-			clientProtocol:                     "http",
-			httpMethod:                         "POST",
-			method:                             "/v1/shelves/100/books?key=api-key",
-			bodyBytes:                          []byte(`{"id": 4, "price_in_usd": 100}`),
-			transcodingPreserveProtoFieldNames: true,
-			wantResp:                           `{"id":"4","priceInUsd":100}`,
-		},
-	}
-	for _, tc := range tests {
-		func() {
-			args := []string{"--service_config_id=" + configID,
-				"--rollout_strategy=fixed"}
-
-			if tc.transcodingAlwaysPrintPrimitiveFields {
-				args = append(args, "--transcoding_always_print_primitive_fields=true")
-			}
-
-			if tc.transcodingAlwaysPrintEnumsAsInts {
-				args = append(args, "--transcoding_always_print_enums_as_ints=true")
-			}
-
-			if tc.transcodingPreserveProtoFieldNames {
-				args = append(args, "--transcoding_preserve_proto_field_names=true")
-			}
-
-			s := env.NewTestEnv(comp.TestTranscodingPrintOptions, platform.GrpcBookstoreSidecar)
-			s.OverrideAuthentication(&confpb.Authentication{
-				Rules: []*confpb.AuthenticationRule{},
-			})
-
-			defer s.TearDown(t)
-			if err := s.Setup(args); err != nil {
-				t.Fatalf("fail to setup test env, %v", err)
-			}
-
-			addr := fmt.Sprintf("localhost:%v", s.Ports().ListenerPort)
-			resp, err := client.MakeHttpCallWithBody(addr, tc.httpMethod, tc.method, "", tc.bodyBytes)
-			if err != nil {
-				t.Errorf("Test (%s): failed with  err %v", tc.desc, err)
-			} else {
-				if !strings.Contains(resp, tc.wantResp) {
-					t.Errorf("Test (%s): failed, expected: %s, got: %s", tc.desc, tc.wantResp, resp)
-				}
-			}
-		}()
-	}
-}
-
-func TestTranscodingIgnoreParameters(t *testing.T) {
-	t.Parallel()
-
-	configID := "test-config-id"
-	type testType struct {
-		desc                                    string
-		clientProtocol                          string
-		httpMethod                              string
-		method                                  string
-		bodyBytes                               []byte
-		wantResp                                string
-		wantError                               string
-		transcodingIgnoreUnknownQueryParameters bool
-		transcodingIgnoreQueryParameters        string
-	}
-
-	tests := []testType{
-		{
-			desc:           "Success. Default setting used to be compared with other test cases.",
-			clientProtocol: "http",
-			httpMethod:     "POST",
-			method:         "/v1/shelves/100/books?key=api-key&unknown_parameter=val",
-			bodyBytes:      []byte(`{"id": 4, "type": 1, "author":"Mark", "price_in_usd": 100}`),
-			wantError:      "503 Service Unavailable",
-		},
-		{
-			desc:                                    "Success. Set transcodingIgnoreUnknownQueryParameters to true.",
-			clientProtocol:                          "http",
-			httpMethod:                              "POST",
-			method:                                  "/v1/shelves/100/books?key=api-key&unknown_parameter_foo=val&unknown_parameter_bar=val",
-			bodyBytes:                               []byte(`{"id": 4, "type": 1, "author":"Mark", "price_in_usd": 100}`),
-			transcodingIgnoreUnknownQueryParameters: true,
-			wantResp:                                `{"id":"4","author":"Mark","type":"COMIC","price_in_usd":100}`,
-		},
-		{
-			desc:                             "Fail. Set transcodingIgnoreQueryParameters with insufficient ignore parameters.",
-			clientProtocol:                   "http",
-			httpMethod:                       "POST",
-			method:                           "/v1/shelves/100/books?key=api-key&unknown_parameter_foo=val&unknown_parameter_bar=val",
-			bodyBytes:                        []byte(`{"id": 4, "type": 1, "author":"Mark", "price_in_usd": 100}`),
-			transcodingIgnoreQueryParameters: "unknown_parameter_foo",
-			wantError:                        "503 Service Unavailable",
-		},
-		{
-			desc:                             "Success. Set right transcodingIgnoreQueryParameters.",
-			clientProtocol:                   "http",
-			httpMethod:                       "POST",
-			method:                           "/v1/shelves/100/books?key=api-key&unknown_parameter_foo=val&unknown_parameter_bar=val",
-			bodyBytes:                        []byte(`{"id": 4, "type": 1, "author":"Mark", "price_in_usd": 100}`),
-			transcodingIgnoreQueryParameters: "unknown_parameter_foo,unknown_parameter_bar",
-			wantResp:                         `{"id":"4","author":"Mark","type":"COMIC","price_in_usd":100}`,
-		},
-	}
-	for _, tc := range tests {
-		func() {
-			args := []string{"--service_config_id=" + configID,
-				"--rollout_strategy=fixed"}
-			if tc.transcodingIgnoreUnknownQueryParameters {
-				args = append(args, "--transcoding_ignore_unknown_query_parameters=true")
-			}
-			if tc.transcodingIgnoreQueryParameters != "" {
-				args = append(args, "--transcoding_ignore_query_parameters="+tc.transcodingIgnoreQueryParameters)
-			}
-
-			s := env.NewTestEnv(comp.TestTranscodingIgnoreQueryParameters, platform.GrpcBookstoreSidecar)
-			s.OverrideAuthentication(&confpb.Authentication{
-				Rules: []*confpb.AuthenticationRule{},
-			})
-
-			defer s.TearDown(t)
-			if err := s.Setup(args); err != nil {
-				t.Fatalf("fail to setup test env, %v", err)
-			}
-
-			addr := fmt.Sprintf("localhost:%v", s.Ports().ListenerPort)
-			resp, err := client.MakeHttpCallWithBody(addr, tc.httpMethod, tc.method, "", tc.bodyBytes)
-			if err == nil {
-				if !strings.Contains(resp, tc.wantResp) {
-					t.Errorf("Test (%s): failed, expected: %s, got: %s", tc.desc, tc.wantResp, resp)
-				}
-				return
-			}
-
-			if tc.wantError == "" {
-				t.Errorf("Test (%s): failed with  err %v", tc.desc, err)
-				return
-			}
-
-			if !strings.Contains(err.Error(), tc.wantError) {
-				t.Errorf("Test (%s): failed with unexpected error, want: %v, get: %s", tc.desc, err, tc.wantError)
-			}
-		}()
 	}
 }
